nvim: changed fterm border to single

+++ .config/nvim/lua/plugins/fterm.lua
@@ -4,7 +4,7 @@ return {
         local fterm = require("FTerm")
 
         fterm.setup({
-            border = "double"
+            border = "single"
         })
 
         local lazygit = fterm:new({
